Bug 645878: Fix windows 32-bit builds (r=fklockii).
This is the error being fixed:

MMgcPortWin.cpp
..\..\AVMPI\MMgcPortWin.cpp(705) : error C2220: warning treated as error
- no 'object' file generated
..\..\AVMPI\MMgcPortWin.cpp(705) : warning C4826: Conversion from 'void
*' to 'unsigned __int64' is sign-extended. This may cause unexpected
runtime behavior.

There was a little bit of email discussion between Stan and Lars about
disabling the warning itself, but it seems easier to me to just put in
the extra cast.  I believe the cast correctly addresses the issue
described by the warning.

+++ AVMPI/MMgcPortWin.cpp
@@ -700,7 +700,7 @@ void AVMPI_freeCodeMemory(void* address, size_t nbytes)
         VMPI_snprintf(buf,
                       sizeof(buf),
                       "AVMPI_freeCodeMemory invariants violated: address=%llu provided=%llu actual=%llu\nAborting.\n",
-                      (unsigned long long)address,
+                      (unsigned long long)(uintptr_t)address,
                       (unsigned long long)nbytes,
                       (unsigned long long)actualBytes);
         VMPI_log(buf);
@@ -728,7 +728,7 @@ void AVMPI_makeCodeMemoryExecutable(void *address, size_t nbytes, bool makeItSo)
         VMPI_snprintf(buf,
                       sizeof(buf),
                       "AVMPI_makeCodeMemoryExecutable invariants violated: address=%llu size=%llu pagesize=%llu\nAborting.\n",
-                      (unsigned long long)address,
+                      (unsigned long long)(uintptr_t)address,
                       (unsigned long long)nbytes,
                       (unsigned long long)pagesize);
         VMPI_log(buf);
